+++ IO/source/MshReader/MshReader3D.cpp
@@ -149,22 +149,22 @@ void MshReader3D::addElements() {
 
 void MshReader3D::defineBoundaryVertices() {
 	for (auto boundary = this->gridData->boundaries.begin(); boundary != this->gridData->boundaries.end(); boundary++) {
-		std::set<long> vertices;
+		std::set<int> vertices;
 		if (boundary->triangleConnectivity.size() > 0) {
 			for (auto j = boundary->triangleConnectivity.cbegin(); j != boundary->triangleConnectivity.cend(); j++) {
 				for (auto k = j->cbegin(); k != j->cend(); k++) {
-					vertices.insert(static_cast<long>(*k));
+					vertices.insert(static_cast<int>(*k));
 				}
 			}
-			boundary->vertices = std::vector<long>(vertices.begin(), vertices.end());
+			boundary->vertices = std::vector<int>(vertices.begin(), vertices.end());
 		}
 		else {
 			for (auto j = boundary->quadrangleConnectivity.cbegin(); j != boundary->quadrangleConnectivity.cend(); j++) {
 				for (auto k = j->cbegin(); k != j->cend(); k++) {
-					vertices.insert(static_cast<long>(*k));
+					vertices.insert(static_cast<int>(*k));
 				}
 			}
-			boundary->vertices = std::vector<long>(vertices.begin(), vertices.end());
+			boundary->vertices = std::vector<int>(vertices.begin(), vertices.end());
 		}
 	}
 }
